add testing code, part4 excercise complete

+++ part4/bloglist-back/test/blog_api.test.js
@@ -5,16 +5,26 @@ const api = supertest(app);
 const testHelper = require('./test_helper');
 const mongoose = require('mongoose');
 const Blog = require('../models/blog');
+const User = require('../models/user');
 
-beforeEach(async () => {
+let authorization = null;
+const loginMocker = async (loginInfo) => {
+  const result = await api.post('/api/login').send(loginInfo);
+  return { authorization: `bearer ${result.body.token}` };
+};
+beforeAll(async () => {
+  jest.setTimeout(30000);
   await Blog.deleteMany({});
-
+  // login in first user, create 4 posts
+  const { username, password } = testHelper.users[0];
+  authorization = await loginMocker({ username, password });
+  const testUser = await User.findOneAndUpdate({ username }, { blogs: [] }, { new: true });
   for (const blog of testHelper.blogs) {
-    const blogDoc = new Blog(blog);
+    const blogDoc = new Blog({ user: testUser._id, ...blog });
     await blogDoc.save();
   }
 });
-describe('what', () => {
+describe('get post', () => {
   test('should get all bloglist from db ', async () => {
     const res = await testHelper.blogsInDb();
     expect(res).toHaveLength(testHelper.blogs.length);
@@ -26,7 +36,7 @@ describe('what', () => {
   });
 });
 
-describe('add data: ', () => {
+describe('add post: ', () => {
   test('should add data correctlys by post request', async () => {
     const newBlog = {
       title: 'Go To Statement Considered Harmful',
@@ -36,6 +46,7 @@ describe('add data: ', () => {
     };
     await api
       .post('/api/blogs')
+      .set(authorization)
       .send(newBlog)
       .expect(200)
       .expect('Content-Type', /application\/json/);
@@ -52,6 +63,7 @@ describe('add data: ', () => {
     };
     await api
       .post('/api/blogs')
+      .set(authorization)
       .send(newBlog)
       .expect(200)
       .expect('Content-Type', /application\/json/);
@@ -64,17 +76,49 @@ describe('add data: ', () => {
       author: 'Edsger W. Dijkstra',
       likes: 3
     };
-    await api.post('/api/blogs').send(newBlog).expect(400);
+    await api.post('/api/blogs').set(authorization).send(newBlog).expect(400);
+  });
+  test('cannot add data without token', async () => {
+    const newBlog = {
+      title: 'no token but want it',
+      author: 'Police should !',
+      url: 'http://www.coronaVirusIsMakeMeSad',
+      likes: 3
+    };
+    await api
+      .post('/api/blogs')
+      .send(newBlog)
+      .expect(401)
+      .expect('Content-Type', /application\/json/)
+      .then((result) => result.body.error.name.includes('unAuthorized'));
   });
 });
 
 describe('delete data', () => {
+  test("cannot delete others'post ", async () => {
+    const { username, password } = testHelper.users[1];
+    const authorizationSecond = await loginMocker({ username, password });
+    const blogs = await testHelper.blogsInDb();
+    const firstBlogId = blogs[0].id;
+    await api.delete(`/api/blogs/${firstBlogId}`).set(authorizationSecond).expect(401);
+  });
+  test('without jwtToken cannot delete', async () => {
+    const blogs = await testHelper.blogsInDb();
+    const firstBlogId = blogs[0].id;
+    try {
+      await api.delete(`/api/blogs/${firstBlogId}`).expect(401);
+    } catch (e) {
+      console.log(e);
+      return;
+    }
+  });
+
   test("delete first blog and check it's removed", async () => {
     const blogs = await testHelper.blogsInDb();
     const firstBlogId = blogs[0].id;
-    await api.delete(`/api/blogs/${firstBlogId}`).expect(204);
+    await api.delete(`/api/blogs/${firstBlogId}`).set(authorization).expect(204);
     const blogsFinal = await testHelper.blogsInDb();
-    expect(blogsFinal).toHaveLength(testHelper.blogs.length - 1);
+    expect(blogsFinal).toHaveLength(blogs.length - 1);
   });
 });
